Add more output information
