+++ PWGHF/Tasks/CMakeLists.txt
@@ -49,14 +49,9 @@ o2physics_add_dpl_workflow(task-pid-studies
                     PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2Physics::EventFilteringUtils O2Physics::AnalysisCCDB
                     COMPONENT_NAME Analysis)
 
-o2physics_add_dpl_workflow(check-mc-pv-contr
-                    SOURCES checkMcPvContr.cxx
+o2physics_add_dpl_workflow(task-mc-injection
+                    SOURCES taskMcInjection.cxx
                     PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2Physics::EventFilteringUtils
-                    COMPONENT_NAME Analysis)
-
-o2physics_add_dpl_workflow(check-aod
-                    SOURCES checkAod.cxx
-                    PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2Physics::EventFilteringUtils O2Physics::AnalysisCCDB
                     COMPONENT_NAME Analysis)
 
 # o2physics_add_dpl_workflow(task-sel-optimisation
